# perf(core): micro-optimize OpsTracker