Relative to #13864

```
Before:
test bench_is_proxy   ... bench:      20,292 ns/iter (+/- 480)
test bench_op_async   ... bench:     385,609 ns/iter (+/- 5,775)
test bench_op_nop     ... bench:      46,042 ns/iter (+/- 1,785)
test bench_op_pi_json ... bench:      49,903 ns/iter (+/- 668)

After:
test bench_is_proxy   ... bench:      20,428 ns/iter (+/- 1,292)
test bench_op_async   ... bench:     386,556 ns/iter (+/- 6,006)
test bench_op_nop     ... bench:      44,772 ns/iter (+/- 1,313)
test bench_op_pi_json ... bench:      48,357 ns/iter (+/- 1,299)
```

Yes, it's only ~1.5ns but that will be 5% of sync-op overhead following #13861